checkpoint macrotasks for MP2, CC2 and LRCC2

+++ src/madness/chem/CCPotentials.cc
@@ -700,12 +700,8 @@ CCPotentials::make_constant_part_macrotask(World& world, const CCPair& pair,
 
     // t1-transformed orbitals
     CC_vecfunction t(MIXED);
-    if (targetstate==CT_CC2 or targetstate==CT_LRCC2)
-    {
-        for (int i=0; i<info.mo_ket.size(); i++) {
-            CCFunction<double,3> t1(info.mo_ket[i] + gs_singles(i).function, i, MIXED);
-            t.insert(i, t1);
-        }
+    if (targetstate==CT_CC2 or targetstate==CT_LRCC2) {
+        t=CCPotentials::make_full_t_intermediate(gs_singles,info);
     }
 
     // construct the projectors
@@ -748,6 +744,24 @@ CCPotentials::make_constant_part_macrotask(World& world, const CCPair& pair,
         return result;
     };
 
+    auto GG = BSHOperator<6>(world, sqrt(-2.0 * pair.bsh_eps), parameters.lo(), parameters.thresh_bsh_6D());
+    GG.destructive() = true;
+    GG.print_timings=false;
+    auto apply_G_and_print = [&](const std::vector<CCPairFunction<double,6>>& cc, std::string name) {
+        std::vector<CCPairFunction<double,6>> tmp1;
+        print("cc in apply_G_and_print:",name,cc.size());
+        for (const auto& tt : cc) tmp1 += GG(copy(tt));
+        print("tmp1 after apply G");
+        for (const auto& tt : tmp1) {
+            print(tt.name());
+            tt.print_size();
+        }
+        tmp1=consolidate(tmp1);
+        tmp1=-2.0*tmp1;
+        MADNESS_CHECK(tmp1.size()==1);
+        tmp1[0].get_function().print_size(name);
+    };
+
     // compute all 6d potentials without applying the SO projector
     std::vector<CCPairFunction<double,6>> V;
     if (targetstate==CT_MP2) {
@@ -761,20 +775,85 @@ CCPotentials::make_constant_part_macrotask(World& world, const CCPair& pair,
     } else if (targetstate==CT_LRCC2) {     // Eq. (25) of Kottmann, JCTC 13, 5956 (2017)
         // eq. (25) Q12t (g~ - omega f12) (|x_i t_j> + |t_i x_j> )
         // note the term omega f12 is included in the reduced_Fock term, see eq. (34)
-        std::vector<std::string> argument={"Ue","KffK","comm_F_Qt_f12","reduced_Fock"};
-        auto Vreg=apply_Vreg(world,t(i),x(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
-        Vreg+=apply_Vreg(world,x(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
-        V=consolidate(apply_in_separated_form(Q12t,Vreg));
+        if (0)
+        {
+            std::vector<std::string> argument={"Ue","KffK","comm_F_Qt_f12","reduced_Fock"};
+            auto Vreg=apply_Vreg(world,x(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            //debug Vreg+=apply_Vreg(world,t(i),x(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            V=consolidate(apply_in_separated_form(Q12t,Vreg));
+            apply_G_and_print(V,"functional response");
+        }
+
+        if (0) {
+            std::vector<std::string> argument={"Ue","KffK","reduced_Fock"};
+            auto Vreg=apply_Vreg(world,x(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            Vreg+=apply_Vreg(world,t(i),x(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+
+            V=consolidate(apply_in_separated_form(Q12t,Vreg));
+            apply_G_and_print(V,"functional response in old terminology, V -- separated");
+//            print("debug 1", V.size());
+//            for (auto& vv : V) {
+//                print(vv.name());
+//                vv.print_size();
+//            }
+//            for (auto& vv : V) vv.convert_to_pure_no_op_inplace();
+//            print("debug 2", V.size());
+//            for (auto& vv : V) {
+//                print(vv.name());
+//                vv.print_size();
+//            }
+//            V=consolidate(V);
+//            print("debug 3", V.size());
+//            for (auto& vv : V) {
+//                print(vv.name());
+//                vv.print_size();
+//            }
+//            print("V separate-consolidate",V.size());
+//            V[0].get_function().print_size("Q12t_FQtQtF_f12.size() separate-consolidate");
+//            apply_G_and_print(V,"functional response in old terminology, V -- consolidated");
+//
+//            auto Q12V=Q12t(Vreg);
+//            print("Q12t_Ue_KffK_redF.size()",Q12V.size());
+//            Q12V[0].get_function().print_size("Q12t_FQtQtF_f12");
+//            apply_G_and_print(Q12V,"functional response in old terminology, Q12V- single term");
+        }
+        if (1) {
+            print_header3("[F12,Qt] f12 |x_i t_j + t_i x_j>");
+            std::vector<std::string> argument={"comm_F_Qt_f12"};
+            auto Vreg=apply_Vreg(world,x(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            Vreg+=apply_Vreg(world,t(i),x(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            auto Q12V=Q12t(Vreg);
+            print("Q12t_FQtQtF_f12.size()",Q12V.size());
+            apply_G_and_print(Q12V,"commutator response in old terminology: Q12V direct");
+
+            V=consolidate(apply_in_separated_form(Q12t,Vreg));
+            apply_G_and_print(V,"commutator response in old terminology: Q12V separated");
+
+        }
 
         // eq. (29) first term: dQt g~ |t_i t_j>
-        argument={"Ue","KffK","comm_F_Qt_f12","reduced_Fock"};
-        auto Vreg1=apply_Vreg(world,t(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
-        V-=consolidate(dQt_1(Vreg1) + dQt_2(Vreg1));
+        if (1) {
+            print_header3("dQt g~ |t_i t_j>");
+            const std::vector<std::string> argument={"Ue","KffK","comm_F_Qt_f12","reduced_Fock"};
+            auto Vreg1=apply_Vreg(world,t(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            V-=consolidate(dQt_1(Vreg1) + dQt_2(Vreg1));
+            apply_G_and_print(consolidate(dQt_1(Vreg1) + dQt_2(Vreg1)),"projector response");
+        }
+
+
 
         // eq. (29) second term = eq. (31): [F12, dQt] f12 |t_i t_j> + omega dQ12t f12 |t_i t_j>
-        argument={"comm_F_dQt_f12"};
-        V+=apply_Vreg(world,t(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+        if (1) {
+            print_header3("[F12, dQt] f12 |t_i t_j>");
+            const std::vector<std::string> argument={"comm_F_dQt_f12"};
+            // V+=apply_Vreg(world,t(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            auto tmp1=apply_Vreg(world,t(i),t(j),gs_singles,ex_singles,info,argument,pair.bsh_eps);
+            apply_G_and_print(tmp1,"commutator projector response");
+        }
+
     }
+    throw;
+
     V=consolidate(V);
     MADNESS_CHECK(V.size()==2);     // term 1: 6d, hi-rank, local; term 2: 3d, low-rank, delocalized
 
@@ -1409,12 +1488,14 @@ CCPotentials::make_constant_part_cc2_Qt_ex(const CCPair& u, const CC_vecfunction
     CCTimer time_cpr(world, "Commutator-Projector Response");
     // Make functional response part: G(QtVreg|xitj + tixj>)
     real_function_6d functional_response;
-    {
+    if (0) {
         time_fr.start();
         const real_function_6d Vxt = (apply_Vreg(xi, tj, Gscreen)).truncate().reduce_rank();
         if (symmetric) {
             real_function_6d V = apply_Q12t(Vxt, t);
+            V.print_size("Q12tVreg");
             const real_function_6d tmp = -2.0 * G(V);
+            tmp.print_size("G(Q12tVreg)");
             functional_response = tmp + swap_particles(tmp);
         } else {
             const real_function_6d Vtx = apply_Vreg(ti, xj, Gscreen);
@@ -1423,22 +1504,28 @@ CCPotentials::make_constant_part_cc2_Qt_ex(const CCPair& u, const CC_vecfunction
             functional_response = -2.0 * G(V);
         }
         time_fr.stop();
-    }     // make Projector Response: -G(OxQt+QtOx)Vreg|titj>
+    }
+    functional_response.print_size("G functional response");
+
+    // make Projector Response: -G(OxQt+QtOx)Vreg|titj>
     real_function_6d projector_response;
-    {
+    if (0) {
         time_pr.start();
         // here is an inconsistency: The Vreg potential will apply (F12-eij) to the |titj> state but we have here (F12-eij-omega)
         // in the future this part here is supposed to be entirely 3D and not use the 6D apply_Vreg function, so right now this is a workaround
         // however, we have to add the missing -omega|titj>
         real_function_6d Vtt_tmp = apply_Vreg(ti, tj, Gscreen);
         real_function_6d titj = make_f_xy(ti, tj);
-        Vtt_tmp = Vtt_tmp - x.omega * titj;
+        print("skipping omega term 1");
+        // Vtt_tmp = Vtt_tmp - x.omega * titj;
         CCPairFunction<double,6> Vtt(Vtt_tmp);
         real_function_6d tmp1;
         real_function_6d tmp2;
         {
             CCPairFunction<double,6> Ox = apply_Ot(Vtt, x, 1);
             CCPairFunction<double,6> OxQt = apply_Qt(Ox, t, 2);
+            OxQt.convert_to_pure_no_op_inplace();
+            OxQt.get_function().print_size("Q12t_FQtQtF_f12");
             tmp1 = -2.0 * apply_G(OxQt, G);
         }
         if (symmetric) tmp2 = swap_particles(tmp1);
@@ -1450,9 +1537,12 @@ CCPotentials::make_constant_part_cc2_Qt_ex(const CCPair& u, const CC_vecfunction
         projector_response = tmp1 + tmp2;
         time_pr.stop();
     }
+    projector_response.print_size("G projector response");
+
     // make commutator response: [F12,Qt12]f12|xitj+tixj> = (O1VQ2t + Q1tO2V)f12|xitj+tixj>
     real_function_6d commutator_response;
     {
+        print_header3("[F12,Qt12]f12|xitj+tixj> = (Ov Qt + Qt Ov) f12 |xitj+tixj>");
         time_cr.start();
         real_function_6d part1;     // the xt parts
         const vector_real_function_3d Vtmp = get_potentials(tau, POT_singles_);
@@ -1482,6 +1572,8 @@ CCPotentials::make_constant_part_cc2_Qt_ex(const CCPair& u, const CC_vecfunction
         commutator_response = part1 + part2;
         time_cr.stop();
     }
+    commutator_response.print_size("G commutator response");
+
     // make Commutator Projector response Response: [F,d/dtau(Qt)] part of d/dtau{([F,Qt])f12|xitj + tixj>}
     // {-O1x[F,Q2t] - Q1t[F,O2x] - [F,O1x]Q2t - [F,Q1t]O2x , used d/dtau(Qt) = -Ox
     //  O1x[F,O2t] - Q1t[F,O2x] - [F,O1x]Q2t + [F,O1t]O2x ,  used [F,Qt] = -[F,Ot]
@@ -1492,10 +1584,17 @@ CCPotentials::make_constant_part_cc2_Qt_ex(const CCPair& u, const CC_vecfunction
     // }f12|titj>
     real_function_6d commutator_projector_response;
     {
+        print_header3("[F12,dQt] f12 |t_i t_j> = (Ox OVt + Qt OVx) f12 |t_i t_j>");
         time_cpr.start();
-        const vector_real_function_3d Vxtmp = sub(world, get_potentials(x, POT_singles_),
-                                                  x.omega * x.get_vecfunction());
+        print("skipping omega term 2");
+        // const vector_real_function_3d Vxtmp = sub(world, get_potentials(x, POT_singles_),
+                                                  // x.omega * x.get_vecfunction());
+        const vector_real_function_3d Vxtmp = get_potentials(x, POT_singles_);
         const vector_real_function_3d Vttmp = get_potentials(tau, POT_singles_);
+
+    madness::print_size(world,Vxtmp,"Vxtmp");
+    madness::print_size(world,Vttmp,"Vttmp");
+    madness::print_size(world,get_active_mo_bra(),"active bra");
         const CC_vecfunction Vx(Vxtmp, UNDEFINED, parameters.freeze());
         const CC_vecfunction Vt(Vttmp, UNDEFINED, parameters.freeze());
         CCPairFunction<double,6> ftt(f12, ti, tj);
@@ -1523,9 +1622,13 @@ CCPotentials::make_constant_part_cc2_Qt_ex(const CCPair& u, const CC_vecfunction
         commutator_projector_response = tmp1 + tmp2;
         time_cpr.stop();
     }
+    commutator_projector_response.print_size("G commutator projector response");
+    print_header3("add all up");
     real_function_6d result =
             functional_response - projector_response + commutator_response + commutator_projector_response;
+    result.print_size("result");
     result = apply_Q12t(result, mo_ket_);
+    result.print_size("Q12t result");
     output.section("Constant Term for Pair " + u.name() + " ended");
     time_fr.info(true, functional_response.norm2());
     time_pr.info(true, projector_response.norm2());
@@ -1615,9 +1718,12 @@ std::vector<CCPairFunction<double,6>>
         V_lowrank += apply_commutator_F_dQt_f12(world,ti,tj,gs_singles,ex_singles,info,&Gscreen);
     }
     V.truncate().reduce_rank();
-    if (parameters.debug()) V.print_size("Vreg");
+    if (parameters.debug()) {
+        V.print_size("Vreg -- pure component");
+        print("V_lowrank.size()",V_lowrank.size());
+    }
 
-    std::vector<CCPairFunction<double,6>> result;
+    std::vector<CCPairFunction<double, 6>> result;
     result+=CCPairFunction<double,6>(V);
     result+=V_lowrank;
     return result;
@@ -2015,8 +2121,8 @@ CCPotentials::apply_commutator_F_Qt_f12(World& world, const CCFunction<double,3>
     auto ftt=std::vector<CCPairFunction<double,6>>({CCPairFunction<double,6>(f12, phi_i.function, phi_j.function)});
 
     const vector_real_function_3d Vtau=info.intermediate_potentials(gs_singles, POT_singles_);
-    Projector<double,3> OVtau(info.mo_bra,Vtau);
-    QProjector<double,3> Qt(world,info.mo_bra,gs_singles.get_vecfunction());
+    Projector<double,3> OVtau(info.get_active_mo_bra(),Vtau);
+    QProjector<double,3> Qt(world,info.get_active_mo_bra(),gs_singles.get_vecfunction());
 
     auto p1=outer(OVtau,Qt);
     auto p2=outer(Qt,OVtau);
@@ -2040,16 +2146,20 @@ CCPotentials::apply_commutator_F_dQt_f12(World& world, const CCFunction<double,3
                                                   const Info& info, const real_convolution_6d *Gscreen) {
     const auto& parameters=info.parameters;
 
-
     auto f12=CCConvolutionOperatorPtr<double,3>(world,OT_F12,parameters);
     auto ftt=std::vector<CCPairFunction<double,6>>({CCPairFunction<double,6>(f12, phi_i.function, phi_j.function)});
 
     const vector_real_function_3d Vtau=info.intermediate_potentials(gs_singles, POT_singles_);
     const vector_real_function_3d Vx=info.intermediate_potentials(ex_singles, POT_singles_);
-    Projector<double,3> OVtau(info.mo_bra,Vtau);
-    Projector<double,3> Ox(info.mo_bra,ex_singles.get_vecfunction());
-    Projector<double,3> OVx(info.mo_bra,Vx);
-    QProjector<double,3> Qt(world,info.mo_bra,gs_singles.get_vecfunction());
+    auto bra=info.get_active_mo_bra();
+    madness::print_size(world,Vtau,"Vtau");
+    madness::print_size(world,Vx,"x");
+    madness::print_size(world,bra,"active bra");
+
+    Projector<double,3> OVtau(bra,Vtau);
+    Projector<double,3> Ox(bra,ex_singles.get_vecfunction());
+    Projector<double,3> OVx(bra,Vx);
+    QProjector<double,3> Qt(world,bra,gs_singles.get_vecfunction());
 
     auto p1_1=outer(OVx,Qt);
     auto p1_2=outer(Qt,OVx);
@@ -2057,7 +2167,7 @@ CCPotentials::apply_commutator_F_dQt_f12(World& world, const CCFunction<double,3
     auto p2_2=outer(OVtau,Ox);
 
     auto result=p1_1(ftt) + p1_2(ftt) - p2_1(ftt) - p2_2(ftt);
-    result=consolidate(result,{""});     // will collect similar terms only
+    result=consolidate(result);     // will collect similar terms only
     MADNESS_CHECK_THROW(result.size()==1 and result[0].is_decomposed(),"apply_Fock_commutator should return a single CCPairFunction");
     return result[0];
 }
@@ -3535,6 +3645,26 @@ CC_vecfunction CCPotentials::make_full_t_intermediate(const CC_vecfunction& tau)
     }
     return result;
 }
+
+/// makes the t intermediates
+/// t_i = mo_ket_(i) + factor*tau(i)
+/// if the core is frozen the core ti will just be mo_ket_
+CC_vecfunction CCPotentials::make_full_t_intermediate(const CC_vecfunction& tau, const Info& info) {
+
+    if (tau.type == HOLE or tau.size()==0) return CC_vecfunction(info.mo_ket,HOLE);
+
+    CC_vecfunction result(MIXED);
+    for (size_t i = 0; i < info.mo_ket.size(); i++) {
+        if (int(i) < info.parameters.freeze()) {
+            result.insert(i, CCFunction<double,3>(info.mo_ket[i],i,MIXED));
+        } else {
+            CCFunction<double,3> t(info.mo_ket[i] + tau(i).function, i, MIXED);
+            result.insert(i, t);
+        }
+    }
+    return result;
+}
+
 
 /// makes the t intermediates
 /// t_i = mo_ket_(i) + tau
